+++ market-contract/src/nft_callbacks.rs
@@ -5,14 +5,14 @@ use crate::*;
 /*
     trait that will be used as the callback from the NFT contract. When nft_approve is
     called, it will fire a cross contract call to this marketplace and this is the function
-    that is invoked. 
+    that is invoked.
 */
 trait NonFungibleTokenApprovalsReceiver {
     fn nft_on_approve(
         &mut self,
         token_id: TokenId,
         owner_id: AccountId,
-        approval_id: u32,
+        approval_id: u64,
         msg: String,
     );
 }
@@ -24,7 +24,7 @@ impl NonFungibleTokenApprovalsReceiver for Contract {
         &mut self,
         token_id: TokenId,
         owner_id: AccountId,
-        approval_id: u32,
+        approval_id: u64,
         msg: String,
     ) {
         /*
