+++ nft-series/src/metadata.rs
@@ -46,9 +46,9 @@ pub struct Token {
     //owner of the token
     pub owner_id: AccountId,
     //list of approved account IDs that have access to transfer the token. This maps an account ID to an approval ID
-    pub approved_account_ids: HashMap<AccountId, u32>,
+    pub approved_account_ids: HashMap<AccountId, u64>,
     //the next approval ID to give out.
-    pub next_approval_id: u32,
+    pub next_approval_id: u64,
 }
 
 //The Json token is what will be returned from view calls.
@@ -64,7 +64,7 @@ pub struct JsonToken {
     //token metadata
     pub metadata: TokenMetadata,
     //list of approved account IDs that have access to transfer the token. This maps an account ID to an approval ID
-    pub approved_account_ids: HashMap<AccountId, u32>,
+    pub approved_account_ids: HashMap<AccountId, u64>,
     //keep track of the royalty percentages for the token in a hash map
     pub royalty: Option<HashMap<AccountId, u32>>,
 }
